anti_ai_club/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.47
Flask-OAuthlib==0.9.6
Flask-Login==0.6.1
Werkzeug==2.0.3
